requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-pymongo==2.3.0
pymongo==4.6.1
PyJWT==2.8.0
Werkzeug==3.0.1
python-dotenv==1.0.1
bcrypt==4.1.2
gunicorn==21.2.0
dnspython==2.5.0